fix: MediaUploader, MediaFileProvider를 domain의 util로 이동, 인터페이스 분리
