+++ android/domain/src/main/kotlin/com/andlife/domain/model/MediaFile.kt
@@ -1,6 +1,4 @@
-package com.andlife.data.util.media
-
-import com.andlife.domain.model.MediaType
+package com.andlife.domain.model
 
 data class MediaFile(
     val uriString: String,
