+++ android/domain/src/main/kotlin/com/andlife/domain/util/media/MediaUploader.kt
@@ -1,6 +1,7 @@
-package com.andlife.data.util.media
+package com.andlife.domain.util.media
 
 import com.andlife.domain.error.DataError
+import com.andlife.domain.model.MediaFile
 import com.andlife.domain.util.Result
 
 interface MediaUploader {
